Chapter 12 From Book Without Exercises
I was having a problem with core data.

When I went to add the table view controller class file, creation was a little different from what the book said. According to page 232, there should have been an option to check "Target for MyContactList Core Data", but there was none. There was just a blank space next to "Targets".

I only have access to chapters 12 and 13 of the book right now (no access to chapter 11), and I'm too busy to spend a super long time trying to figure out what's wrong. So I am just going to use chapter 12 from the book's Github and do the exercises on that.