Chapter 14
Since the emulator won't allow you to get permission to use the camera, I have not really been able to test. However, I did try to do exercises 2 and 3, though since it seems like the emulators are supposed to have a camera, I technically can't test those either. But assuming that the book is correct about how to check for the camera, they should work.